# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API Documentation",
        "description": "Dokumentasi API untuk aplikasi E-Ticketing Helpdesk berbasis Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Menampilkan daftar tiket",
                "description": "Mengambil daftar tiket helpdesk milik user atau seluruh tiket untuk admin/helpdesk.",
                "operationId": "ddc3a5e66a83482e874c84e71ebc4505",
                "responses": {
                    "200": {
                        "description": "Daftar tiket berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Membuat tiket baru",
                "description": "User membuat tiket helpdesk baru.",
                "operationId": "2a7d76295d0f9f3e0e63adb2ed4a0596",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laptop tidak bisa menyala"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Laptop mati total setelah update Windows."
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Hardware"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "High"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tiket berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Menampilkan detail tiket",
                "description": "Mengambil detail tiket berdasarkan ID.",
                "operationId": "da5292e25da44db69ed981a43357753c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail tiket berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}/status": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Mengubah status tiket",
                "description": "Mengubah status tiket helpdesk, misalnya open, in_progress, resolved, atau closed.",
                "operationId": "67aa843d6c1b3c738a0a1d80c45cf7bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "in_progress"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status tiket berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}/assign": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Assign tiket ke helpdesk",
                "description": "Menugaskan tiket kepada helpdesk tertentu.",
                "operationId": "4e73265b6b81db806d8e5584c0e8acb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigned_to"
                                ],
                                "properties": {
                                    "assigned_to": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiket berhasil di-assign"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}/history": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Menampilkan riwayat tiket",
                "description": "Mengambil riwayat perubahan status atau aktivitas pada tiket.",
                "operationId": "0a23d942b12f07e3d91803d79fc2db5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tiket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat tiket berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}